feat: Remove settings, move delegation bits to client
Clean up CI, switch to NPM for ci.
Misc cleanup.
Remove prettier plugin.
Fix typecheck / api build.
