+++ .github/workflows/ci.yml
@@ -19,7 +19,7 @@ jobs:
           node-version: 16
       - name: Run Tests
         run: |
-          yarn install --immutable --immutable-cache --check-cache
+          npm ci
           npm test
       - name: Build
         run: npm run build
