+++ .github/workflows/pr.yml
@@ -13,5 +13,5 @@ jobs:
         uses: actions/setup-node@v3
         with:
           node-version: 16
-      - run: yarn install --immutable --immutable-cache --check-cache
+      - run: npm ci
       - run: npm test
