+++ .jsdoc.json
@@ -1,0 +1,3 @@
+{
+  "plugins": ["node_modules/jsdoc-tsimport-plugin/index.js"]
+}
