+++ .prettierrc.cjs
@@ -2,7 +2,4 @@ module.exports = {
   trailingComma: 'es5',
   semi: false,
   singleQuote: true,
-  importOrder: ['^[./]'],
-  importOrderSeparation: true,
-  importOrderSortSpecifiers: true,
 }
