+++ LICENSE.md
@@ -230,4 +230,3 @@ THE SOFTWARE.
 ```
 
 </details>
-
