+++ package.json
@@ -14,12 +14,14 @@
     "dist"
   ],
   "scripts": {
-    "build": "npm run typecheck",
-    "build:api_docs": "jsdoc2md -f src/index.js > API.md",
-    "x-pretest": "npm run prettier && npm run typecheck",
+    "build": "npm run build:api_docs && npm run build:types",
+    "build:api_docs": "jsdoc2md -c .jsdoc.json -f src/index.js > API.md",
+    "build:types": "tsc --declaration --emitDeclarationOnly --declarationDir ./dist",
+    "pretest": "npm run prettier && npm run typecheck",
     "test": "vitest run",
-    "typecheck": "tsc",
+    "typecheck": "tsc --noEmit",
     "prettier": "prettier -c '{src,patches,test}/**/*.{js,ts,yml,json}' --ignore-path .gitignore",
+    "prettier:fix": "prettier --ignore-unknown --write --ignore-path .gitignore .",
     "test:coverage": "vitest run --coverage",
     "test:dev": "vitest -w"
   },
@@ -37,10 +39,10 @@
     "uint8arrays": "^4.0.2"
   },
   "devDependencies": {
-    "@trivago/prettier-plugin-sort-imports": "^3.3.0",
     "@vitest/coverage-c8": "^0.23.4",
     "ipjs": "^5.2.0",
     "jsdoc-to-markdown": "^7.1.1",
+    "jsdoc-tsimport-plugin": "^1.0.5",
     "lint-staged": "^13.0.3",
     "prettier": "^2.7.1",
     "semantic-release": "^19.0.5",
