+++ src/delegation/generation.js
