+++ tsconfig.json
@@ -8,11 +8,11 @@
     "target": "ES2022",
     "module": "ES2022",
     "lib": ["ES2022", "DOM", "DOM.Iterable"],
-    "noEmit": false,
-    "noEmitOnError": true,
-    "emitDeclarationOnly": true,
+    // "noEmit": false,
+    // "noEmitOnError": true,
+    // "emitDeclarationOnly": true,
     "sourceMap": true,
-    "declaration": true,
+    // "declaration": true,
     "declarationMap": true,
     "incremental": true,
     "composite": true,
